backport of synchronous logging, COUCHDB-761.  Patch by Randall Leeds.
This patch fixes a bug where calls to get the log level could time
out, resulting in an unexpected {ref(), integer()} message in the
client's mailbox.  This would crash some gen_servers that did not
expect the message.

This backport also includes r954240, the hack for logging during
make check.

git-svn-id: https://svn.apache.org/repos/asf/couchdb/branches/0.10.x@954242 13f79535-47bb-0310-9956-ffa450edef68