Add `make dev-test` target for test suite.
This makes the erlang sources compliled with

    `erlc +export_all`

so that unexported functions can be tested alongside exported
ones.

This isn’t entirely baked, as I call `make clean` after `make
check` to avoid `dist`-ing binaries with +export_all.

Robert Newson suggests to have the sources compiled with
+export_all to separate location and run `make dev-text` from
that directory, which should be cleaner. Consider this a TBD.